dv360-reports-automation/dv360_reports_create/requirements.txt
google-cloud==0.34.0
pandas==1.3.5
requests==2.28.0
google-cloud-bigquery==3.2.0
google-cloud-storage==2.4.0
oauth2client
google-api-python-client
db-dtypes==1.0.3